refactor: SignIn 화면을 Android Architecture Samples 패턴으로 리팩토링
